Add Role in Member model instead of User (#69)
Currently role is part of firebase user customClaims field. Changing the role doesn't immediately reflect the changes in UI. Reason is that customClaims in the UI gets updated after token expires. Only option is to forcefully refresh token or logout and login once again to see those changes. So, instead of relying on firebase we want to manage Roles in our own DB in the members model.

 Update Member model in prisma schema
 Update getAllMembers service to return Role from member
 Update Roles decorator to check role from member
 Update service for /updateRole endpoint
 Update createworkspace service to add role in member's field
 Change FE to use role from member instead of user

Also, added hyperlink to career site in org settings page